+++ js/init.js
@@ -186,11 +186,11 @@
     $('.materialboxed').materialbox();
     $('.modal').modal();
     $('.scrollspy').scrollSpy();
-    $('.button-collapse').sideNav({'edge': 'left'});
     $('.datepicker').pickadate({selectYears: 20});
     $('.timepicker').pickatime();
     $('.tooltipped').tooltip();
     $('select').not('.disabled').material_select();
+    $('.sidenav').sidenav();
     $('input.autocomplete').autocomplete({
       data: {"Apple": null, "Microsoft": null, "Google": 'http://placehold.it/250x250'},
     });